+++ agent_ai/agent_ai.py
@@ -35,8 +35,8 @@ import numpy as np
 import requests
 
 # Local application imports
-from .utils.logger import get_logger
-from .validation.validator import (
+from agent_ai.utils.logger import get_logger
+from agent_ai.validation.validator import (
     check_missing_dependencies,
     validate_module_registration,
 )
