+++ watchers/watchers_tools/solenoid_watcher/controller/solenoid_controller.py
@@ -6,7 +6,7 @@ Este módulo implementa un controlador PID para ajustar la señal de control
 del solenoide, utilizando el modelo físico implementado en solenoid_model.
 """
 
-from .agent_ai.model.solenoid_model import simulate_solenoid
+from watchers.watchers_tools.solenoid_watcher.model.solenoid_model import simulate_solenoid
 
 
 class SolenoidController:
